Worked on animation controls, testing materials, added wolf animations, added animation script
worked on the animations for the squirrel and going to continue working on them to tie them to his velocity instead of his direction, have also added some materials and experimented with decals in my testing area. pushing this now so i can experiment in the animation script without causing issues later on.