modifications to level 5, added glow slight Ratatoskr
edited the room to fit with the brown colour scheme, set up the map to use the prefabs, edited ratatoskr prefab to have a light underneath so that you can see the floor beneath him. will need to remove a few lights to fix framerate when looking at the entrance.